Fix map loading state
Also separates SVG rendering from loading state handling and adds a fade-in
animation to the map appearing (like all other IndexBreakdowns)
